+++ code/cell.modules/pkg.sys/ui.dev/src/types/Actions/ActionPanel.ts
@@ -1,4 +1,4 @@
-import { t } from './common';
+import { t } from '../common';
 
 export type ActionPanelProps = {
   scrollable?: boolean;
