+++ code/cell.modules/pkg.sys/ui.dev/src/types/Actions/Actions.builder.ts
@@ -1,4 +1,4 @@
-import { t } from './common';
+import { t } from '../common';
 
 type O = Record<string, unknown>;
