+++ code/cell.modules/pkg.sys/ui.dev/src/types/Actions/Actions.handler.ts
@@ -1,4 +1,4 @@
-import { t } from './common';
+import { t } from '../common';
 
 /**
  * Handler that retrieves the current context.
