+++ code/cell.modules/pkg.sys/ui.dev/src/types/Actions/Actions.model.item.select.ts
@@ -1,4 +1,4 @@
-import { t } from './common';
+import { t } from '../common';
 
 /**
  * INPUT: A button with a toggle switch (boolean)
