+++ code/cell.modules/pkg.sys/ui.dev/src/types/index.ts
@@ -1,15 +1,7 @@
 export * from './types';
 export * from './types.event';
 
-export * from './types.ActionPanel';
-export * from './types.Actions.builder';
-export * from './types.Actions.event';
-export * from './types.Actions.factory';
-export * from './types.Actions.handler';
-export * from './types.Actions.model.item';
-export * from './types.Actions.model.item.select';
-export * from './types.Actions.model';
-export * from './types.Actions.subject';
-
 export * from './types.Host';
 export * from './types.Hosted';
+
+export * from './Actions';
